Crate robonomics_runtime

The Robonomics runtime. This can be compiled with #[no_std], ready for Wasm.

Modules

api

constants
A set of constant values used in substrate runtime.

impls
Some configurable implementations as associated type for the substrate runtime.

Structs

AdjustmentVariable

AvailableBlockRatio

BasicDeposit

BlockHashCount

BondingDuration

Burn

CandidacyBond

CouncilMaxProposals

CouncilMotionDuration

CreationFee

DealWithFees

DepositBase

DepositFactor

DesiredMembers

DesiredRunnersUp

DisabledValidatorsThreshold

ElectionLookahead

ElectionsPhragmenModuleId

EpochDuration

ExistentialDeposit

ExpectedBlockTime

FieldDeposit

GenesisConfig

ImOnlineUnsignedPriority

IndexDeposit

MaxAdditionalFields

MaxIterations

MaxNominatorRewardedPerValidator

MaxRegistrars

MaxSignatories

MaxSubAccounts

MaximumBlockLength

MaximumBlockWeight
We allow for 2 seconds of compute with a 6 second average block time.

MaximumExtrinsicWeight
Assume 10% of weight for average on_initialize calls

MaximumSchedulerWeight

MinSolutionScoreBump

MinimumMultiplier

MinimumPeriod

ModuleToIndex
Provides an implementation of ModuleToIndex to map a module to its index in the runtime.

OffencesWeightSoftLimit

Origin

ProposalBond

ProposalBondMinimum

ReportLatency

RewardCurve

Runtime

RuntimeApi

RuntimeApiImpl
Implements all runtime apis for the client side.

SessionDuration

SessionKeys

SessionsPerEra

SlashDeferDuration

SpendPeriod

StakingUnsignedPriority
We prioritize im-online heartbeats over phragmen solution submission.

SubAccountDeposit

TargetBlockFullness

TermDuration

TipCountdown

TipFindersFee

TipReportDepositBase

TipReportDepositPerByte

TransactionByteFee

TransferFee

TreasuryModuleId

UncleGenerations

Version

VotingBond

WindowSize

Enums

Call

Event

OriginCaller

StakerStatus
Indicates the initial status of the staker.

Constants

VERSION
This runtime version.

WASM_BINARY

WASM_BINARY_BLOATY

Functions

native_version
The version infromation used to identify this runtime when compiled natively.

wasm_binary_unwrap
Wasm binary unwrapped. If built with BUILD_DUMMY_WASM_BINARY, the function panics.

Type Definitions

Address
The address format for describing accounts.

AuthorityDiscovery

AuthorityDiscoveryConfig

Authorship

Babe

BabeConfig

Balances

BalancesConfig

Block
Block type as expected by this runtime.

BlockId
BlockId type as expected by this runtime.

CheckedExtrinsic
Extrinsic type that has already been checked.

Context
The type used as a helper for interpreting the sender of transactions.

Council

CouncilConfig

Datalog

Elections

ElectionsConfig

Executive
Executive: handles dispatch to the various modules.

FinalityTracker

Grandpa

GrandpaConfig

Header
Block header type as expected by this runtime.

Historical

Identity

ImOnline

ImOnlineConfig

Indices

IndicesConfig

Launch

Liability

Multisig

Offences

RandomnessCollectiveFlip

Scheduler

Session

SessionConfig

SignedExtra
The SignedExtension to the basic transaction logic.

SignedPayload
The payload being signed in transactions.

Staking

StakingConfig

Sudo

SudoConfig

System

SystemConfig

Timestamp

TransactionPayment

Treasury

TreasuryConfig

UncheckedExtrinsic
Unchecked extrinsic type as expected by this runtime.

Utility